Security update

+++ pages/views.py
@@ -14,11 +14,13 @@ from django.core.exceptions import ObjectDoesNotExist
 from django.db.models import Q
 from django.views.generic.edit import UpdateView
 from django.core.paginator import Paginator, EmptyPage, PageNotAnInteger
+from django.contrib.auth.decorators import login_required
 
 def index (request):
   return render(request, 'index.html', {})
   #template = loader.get_template('index.html')
   #return HttpResponse(template.render())
+
 
 def registration(request):
     if request.method == 'POST':
@@ -79,38 +81,44 @@ def update_password(request):
     else:
         messages.success(request, "User must be logged in to do this")
         return redirect('loginBCB')
-    
-def music(request):
-    score = Songs.objects.filter(moderated='Yes').order_by('title').values()
-    page_num = request.GET.get('page', 1)
-    paginator = Paginator(score, 5)
-  
-    try:
-        items_page = paginator.page(page_num)
-        items_page_items = items_page.object_list
-    except PageNotAnInteger:
-        items_page = paginator.page(1)
-    except EmptyPage:
-        items_page = paginator.page(paginator.num_pages)
-    #score = Songs.objects.all()
 
-    return render(request, "musicMod.html", { "items_page": items_page})
+
+def music(request):
+    if request.user.is_authenticated:
+        score = Songs.objects.filter(moderated='Yes').order_by('title').values()
+        page_num = request.GET.get('page', 1)
+        paginator = Paginator(score, 5)
+    
+        try:
+            items_page = paginator.page(page_num)
+            items_page_items = items_page.object_list
+        except PageNotAnInteger:
+            items_page = paginator.page(1)
+        except EmptyPage:
+            items_page = paginator.page(paginator.num_pages)
+        #score = Songs.objects.all()
+        return render(request, "musicMod.html", { "items_page": items_page})
+    else:
+        return render(request, "login.html")
 
 def musicMod(request):
-    score = Songs.objects.all().order_by('title').values()
-    page_num = request.GET.get('page', 1)
-    paginator = Paginator(score, 5)
-  
-    try:
-        items_page = paginator.page(page_num)
-        items_page_items = items_page.object_list
-    except PageNotAnInteger:
-        items_page = paginator.page(1)
-    except EmptyPage:
-        items_page = paginator.page(paginator.num_pages)
-    #score = Songs.objects.all()
+    if request.user.is_authenticated:
+        score = Songs.objects.all().order_by('title').values()
+        page_num = request.GET.get('page', 1)
+        paginator = Paginator(score, 5)
+    
+        try:
+            items_page = paginator.page(page_num)
+            items_page_items = items_page.object_list
+        except PageNotAnInteger:
+            items_page = paginator.page(1)
+        except EmptyPage:
+            items_page = paginator.page(paginator.num_pages)
+        #score = Songs.objects.all()
 
-    return render(request, "musicMod.html", { "items_page": items_page})
+        return render(request, "musicMod.html", { "items_page": items_page})
+    else:
+        return render(request, "login.html")
 
 def search_music(request):
     if request.method == 'GET':
@@ -126,38 +134,48 @@ def search_music(request):
 
 
 def add_tune(request):
-    if request.method == 'POST':
-        form = SongsForm(request.POST, request.FILES)
-        print('here')
-        if form.is_valid():
-            form.save()
-            return redirect('music')
+    if request.user.is_authenticated:
+        if request.method == 'POST':
+            form = SongsForm(request.POST, request.FILES)
+            print('here')
+            if form.is_valid():
+                form.save()
+                return redirect('music')
+            else:
+                print('Not Here')
         else:
-            print('Not Here')
+            print('here 2')
+            form = SongsForm()
+        return render(request, 'add_tune.html', {'form':form})
     else:
-        print('here 2')
-        form = SongsForm()
-    return render(request, 'add_tune.html', {'form':form})
+        return render(request, 'login.html')
 
 def delete_tune(request, id):
-    song = get_object_or_404(Songs, id=id)
-    song.delete()
-    return redirect('music')
+    if request.user.is_authenticated:
+        song = get_object_or_404(Songs, id=id)
+        song.delete()
+        return redirect('music')
+    else:
+        return render(request, 'login.html')
 
 
 def download_music(request, name):
-    file = os.path.join(settings.BASE_DIR, 'uploads/songs/',name)
-    with open(file, 'rb') as pdf:
-        response = HttpResponse(pdf.read(), content_type='application/pdf')
-        #response['Content-Disposition'] = 'inline;filename=some_file.pdf'
-        return response
-    pdf.closed
+    if request.user.is_authenticated:
+        file = os.path.join(settings.BASE_DIR, 'uploads/songs/',name)
+        with open(file, 'rb') as pdf:
+            response = HttpResponse(pdf.read(), content_type='application/pdf')
+            #response['Content-Disposition'] = 'inline;filename=some_file.pdf'
+            return response
+        pdf.closed
+    else:
+        return render(request, 'login.html')
 
 
 #def edit_music(request, id):
 
 
 def edit_music(request, id):
+    if request.user.is_authenticated:
         tune = Songs.objects.get(id=id)
         if request.method == 'POST':
             form = SongsForm(request.POST, instance=tune)
@@ -168,6 +186,9 @@ def edit_music(request, id):
             form = SongsForm(instance=tune)
                 
         return render(request, 'edit_music.html',{'form': form})
+    else:
+        return render(request, 'login.html')
+
 
 
 
@@ -184,22 +205,25 @@ def edit_music(request, id):
     #    return render(request, 'edit_music.html', {"edit_score":edit_score})
 
 def update_file(request, id):
-    if request.method == 'POST':
-        file = request.FILES['update_file']
+    if request.user.is_authenticated:
+        if request.method == 'POST':
+            file = request.FILES['update_file']
 
-        file_name = request.FILES['update_file'].name
-        folder = 'songs/'
-        fs = FileSystemStorage()
-        file = fs.save(file.name, file)
-        print(file)
-        fileurl = fs.url(file)
-        report = file_name
+            file_name = request.FILES['update_file'].name
+            folder = 'songs/'
+            fs = FileSystemStorage()
+            file = fs.save(file.name, file)
+            print(file)
+            fileurl = fs.url(file)
+            report = file_name
 
-        Songs.objects.filter(id=id).update(file=file)
+            Songs.objects.filter(id=id).update(file=file)
 
-        return redirect('music')
+            return redirect('music')
+        else:
+            return render(request, 'update_file.html')
     else:
-        return render(request, 'update_file.html')
+        return render(request, 'login.html')
 
 
 def gallery(request):
@@ -208,38 +232,48 @@ def gallery(request):
     return render(request, 'gallery.html', {'articles':articles})
 
 def add_gallery_item(request):
-    if request.method == 'POST':
-        form = GalleryForm(request.POST, request.FILES)
-        print('here')
-        if form.is_valid():
-            form.save()
-            return redirect('gallery')
+    if request.user.is_authenticated:
+        if request.method == 'POST':
+            form = GalleryForm(request.POST, request.FILES)
+            print('here')
+            if form.is_valid():
+                form.save()
+                return redirect('gallery')
+            else:
+                print('Not Here')
         else:
-            print('Not Here')
+            print('here 2')
+            form = GalleryForm()
+        return render(request, 'add_gallery_item.html', {'form':form})
     else:
-        print('here 2')
-        form = GalleryForm()
-    return render(request, 'add_gallery_item.html', {'form':form})
+        return render(request, 'login.html')
 
 def delete_gallery_item(request, id):
-    img = get_object_or_404(Gallery, id=id)
-    img.delete()
-    return redirect('gallery')
+    if request.user.is_authenticated:
+        img = get_object_or_404(Gallery, id=id)
+        img.delete()
+        return redirect('gallery')
+    else:
+        return render(request, 'login.html')
+
 
 
 def add_news_item(request):
-    if request.method == 'POST':
-        form = NewsForm(request.POST, request.FILES)
-        print('here')
-        if form.is_valid():
-            form.save()
-            return redirect('news')
+    if request.user.is_authenticated:
+        if request.method == 'POST':
+            form = NewsForm(request.POST, request.FILES)
+            print('here')
+            if form.is_valid():
+                form.save()
+                return redirect('news')
+            else:
+                print('Not Here')
         else:
-            print('Not Here')
+            print('here 2')
+            form = NewsForm()
+        return render(request, 'add_news_item.html', {'form':form})
     else:
-        print('here 2')
-        form = NewsForm()
-    return render(request, 'add_news_item.html', {'form':form})
+        return render(request, 'login.html')
 
 
 def news(request):
@@ -248,9 +282,12 @@ def news(request):
     return render(request, 'news.html', {'articles':articles})
 
 def delete_news_item(request, id):
-    img = get_object_or_404(News, id=id)
-    img.delete()
-    return redirect('news')
+    if request.user.is_authenticated:
+        img = get_object_or_404(News, id=id)
+        img.delete()
+        return redirect('news')
+    else:
+        return render(request, 'login.html')
 
 
 def links(request):
@@ -259,75 +296,98 @@ def links(request):
     return render(request, 'links.html', {'links':links})
 
 def delete_link(request, id):
-    link = get_object_or_404(Links, id=id)
-    link.delete()
-    return redirect('links')
+    if request.user.is_authenticated:
+        link = get_object_or_404(Links, id=id)
+        link.delete()
+        return redirect('links')
+    else:
+        return render(request, 'login.html')
 
 def add_link(request):
-    if request.method == 'POST':
-        form = LinkForm(request.POST)
-        print('here')
-        if form.is_valid():
-            form.save()
-            return redirect('links')
+    if request.user.is_authenticated:
+        if request.method == 'POST':
+            form = LinkForm(request.POST)
+            print('here')
+            if form.is_valid():
+                form.save()
+                return redirect('links')
+            else:
+                print('Not Here')
         else:
-            print('Not Here')
+            print('here 2')
+            form = LinkForm()
+        return render(request, 'add_link.html', {'form':form})
     else:
-        print('here 2')
-        form = LinkForm()
-    return render(request, 'add_link.html', {'form':form})
+        return render(request, 'login.html')
 
 
 def edit_gallery(request, id):
-    edit_img = Gallery.objects.get(id=id)
-    if request.method == "POST":
-        edit_img.heading= request.POST['heading']
-        edit_img.content_text = request.POST['content_text']
-        edit_img.save()        
-        return redirect('gallery')
-    print('here')
-    edit_img = Gallery.objects.get(id=id)
-
-    return render(request, 'edit_gallery.html', {"edit_img":edit_img})
+    if request.user.is_authenticated:
+        edit_img = Gallery.objects.get(id=id)
+        if request.method == "POST":
+            edit_img.heading= request.POST['heading']
+            edit_img.content_text = request.POST['content_text']
+            edit_img.save()        
+            return redirect('gallery')
+        print('here')
+        edit_img = Gallery.objects.get(id=id)
+        return render(request, 'edit_gallery.html', {"edit_img":edit_img})
+    else:
+        return render(request, 'login.html')
 
 def contact(request):
 
     return render(request, 'contact.html', {})
 
 def documents(request):
-    docs = Documents.objects.all().values()
-    ordering = ['article_created_at']
-    return render(request, 'documents.html', {'docs':docs})
+    if request.user.is_authenticated:
+        docs = Documents.objects.all().values()
+        ordering = ['article_created_at']
+        return render(request, 'documents.html', {'docs':docs})
+    else:
+        return render(request, 'login.html')
 
 def documentMod(request):
-    docs = Documents.objects.all().values()
-    ordering = ['article_created_at']
-    return render(request, 'documentMod.html', {'docs':docs})
-    
+    if request.user.is_authenticated:
+        docs = Documents.objects.all().values()
+        ordering = ['article_created_at']
+        return render(request, 'documentMod.html', {'docs':docs})
+    else:
+        return render(request, 'login.html')
+        
 
 def add_documents(request):
-    if request.method == 'POST':
-        form = DocumentForm(request.POST, request.FILES)
-        print('here')
-        if form.is_valid():
-            form.save()
-            return redirect('documents')
+    if request.user.is_authenticated:
+        if request.method == 'POST':
+            form = DocumentForm(request.POST, request.FILES)
+            print('here')
+            if form.is_valid():
+                form.save()
+                return redirect('documents')
+            else:
+                print('Not Here')
         else:
-            print('Not Here')
+            print('here 2')
+            form = DocumentForm()
+        return render(request, 'add_document.html', {'form':form})
     else:
-        print('here 2')
-        form = DocumentForm()
-    return render(request, 'add_document.html', {'form':form})
+        return render(request, 'login.html')
 
 def download_documents(request, name):
-    file = os.path.join(settings.BASE_DIR, 'upload/documents/',name)
-    with open(file, 'rb') as pdf:
-        response = HttpResponse(pdf.read(), content_type='application/pdf')
-        #response['Content-Disposition'] = 'inline;filename=some_file.pdf'
-        return response
-    pdf.closed
+    if request.user.is_authenticated:
+        file = os.path.join(settings.BASE_DIR, 'upload/documents/',name)
+        with open(file, 'rb') as pdf:
+            response = HttpResponse(pdf.read(), content_type='application/pdf')
+            #response['Content-Disposition'] = 'inline;filename=some_file.pdf'
+            return response
+        pdf.closed
+    else:
+        return render(request, 'login.html')
 
 def delete_document(request, id):
-    docs = get_object_or_404(Documents, id=id)
-    docs.delete()
-    return redirect('documents')
+    if request.user.is_authenticated:
+        docs = get_object_or_404(Documents, id=id)
+        docs.delete()
+        return redirect('documents')
+    else:
+        return render(request, 'login.html')
